backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
chromadb
sentence-transformers
anthropic
slowapi